fix: render edge tooltips above nodes and edges (#82)
Extract tooltip from MapEdge into MapEdgeTooltip component. Render
in MapCanvas as Layer 4 (topmost), after time axis → edges → nodes.
Also make edge tooltips visually lighter (smaller font, less padding)
to be subordinate to future node popovers.

Closes #82
